Feat: Add Folder path to Stable

+++ Episode.cs
@@ -2,7 +2,7 @@ using System;
 class Episode
 {
     public string season { get; set; }
-    public int episodeNumberOverall { get; set; } //EpisodeNumber
+    public float episodeNumberOverall { get; set; } //EpisodeNumber
     public int episodeNumberInSeason { get; set; } //EpisodeNumber2
     public string title { get; set; }
     public string titleRomaji { get; set; }
